APP(BE)/web-server/pyproject.toml
[tool.poetry]
name = "mili-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "mili_project"}]

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.0"
djangorestframework = "3.13.1"
channels = "3.0.4"
channels-redis = "3.3.1"
pytest-asyncio = "0.16.0"
pytest-django = "4.5.2"
Pillow = "8.4.0"
djangorestframework-jwt = "1.11.0"
mysqlclient = "^2.1.1"
daphne = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
